collected some types into common.go. removed sorting methods from Fields. improved godoc.
