--- verbose.go
@@ -1,4 +1,0 @@
-package logng
-
-// Verbose is the type of verbose level.
-type Verbose int
